Unsigned types are unsupported in ImpCode, thus group sizes and IDs must be typecast.

+++ src/Futhark/CodeGen/ImpGen/Kernels/ToOpenGL.hs
@@ -257,15 +257,15 @@ inShaderOperations body =
 
         shaderOps :: GenericC.OpCompiler KernelOp ShaderState
         shaderOps (GetGroupId v i) =
-          GenericC.stm [C.cstm|$id:v = gl_WorkGroupID[$int:i];|]
+          GenericC.stm [C.cstm|$id:v = int32_t(gl_WorkGroupID[$int:i]);|]
         shaderOps (GetLocalId v i) =
-          GenericC.stm [C.cstm|$id:v = gl_LocalInvocationID[$int:i];|]
+          GenericC.stm [C.cstm|$id:v = int32_t(gl_LocalInvocationID[$int:i]);|]
         shaderOps (GetLocalSize v i) =
-          GenericC.stm [C.cstm|$id:v = gl_LocalGroupSizeARB[$int:i];|]
+          GenericC.stm [C.cstm|$id:v = int32_t(gl_LocalGroupSizeARB[$int:i]);|]
         shaderOps (GetGlobalId v i) =
-          GenericC.stm [C.cstm|$id:v = gl_GlobalInvocationID[$int:i];|]
+          GenericC.stm [C.cstm|$id:v = int32_t(gl_GlobalInvocationID[$int:i]);|]
         shaderOps (GetGlobalSize v i) =
-          GenericC.stm [C.cstm|$id:v = gl_NumWorkGroups[$int:i];|]
+          GenericC.stm [C.cstm|$id:v = int32_t(gl_NumWorkGroups[$int:i]);|]
         shaderOps (GetLockstepWidth v) =
           GenericC.stm [C.cstm|$id:v = LOCKSTEP_WIDTH;|]
         shaderOps (Barrier f) = do
